Merge in PUSH/bookingsg-api from feature/BOOKINGSG-1540-labels-categories to master

* commit 'd2ceafac57eef3b6f5b3af254917a3746babb0e1':
  [AL] Reduce
  [AL-WN] SQL improvement
  [AL-WN] SQL improvement
  PR revied
  Lint
  Merge master
  Fix bug + add tests
  Rename database categorie to labelCategorie
  Rename categories to labelCategories
  WIP
  WIP
  pushing changes
  wip added category migration file
  wip added categories to services
  wip
  wip adding service categories

+++ src/components/labels/labels.repository.ts
@@ -1,9 +1,10 @@
 import { Inject, InRequestScope } from 'typescript-ioc';
 import { RepositoryBase } from '../../core/repository';
-import { Label } from '../../models/entities';
+import { Label, LabelCategory, Service } from '../../models/entities';
 import { UserContext } from '../../infrastructure/auth/userContext';
 import { andWhere } from '../../tools/queryConditions';
 import { ServicesQueryAuthVisitor } from '../services/services.auth';
+import { groupByKey } from "../../tools/collections";
 
 @InRequestScope
 export class LabelsRepository extends RepositoryBase<Label> {
@@ -16,7 +17,7 @@ export class LabelsRepository extends RepositoryBase<Label> {
 
 	public async delete(data: Label[]) {
 		const repository = await this.getRepository();
-
+		if (!data.length) return;
 		await repository.delete(data.map((label) => label.id));
 	}
 
@@ -25,8 +26,12 @@ export class LabelsRepository extends RepositoryBase<Label> {
 		return repository.save(data);
 	}
 
-	public async find(options: { serviceId: number; skipAuthorisation?: boolean }): Promise<Label[]> {
-		const { serviceId } = options;
+	public async find(options: {
+		serviceIds?: number[];
+		categoryIds?: number[];
+		skipAuthorisation?: boolean;
+	}): Promise<Label[]> {
+		const { serviceIds, categoryIds } = options;
 		const authGroups = await this.userContext.getAuthGroups();
 
 		const { userCondition, userParams } = options.skipAuthorisation
@@ -34,12 +39,40 @@ export class LabelsRepository extends RepositoryBase<Label> {
 			: await new ServicesQueryAuthVisitor('servicelabel').createUserVisibilityCondition(authGroups);
 
 		const repository = await this.getRepository();
-		const serviceCondition = 'label."_serviceId" = :serviceId';
+		const serviceCondition = serviceIds?.length ? 'label."_serviceId" IN (:...serviceIds) ' : '';
+		const categoryCondition = categoryIds?.length ? 'label."_categoryId" IN  (:...categoryIds) ' : '';
 
 		return repository
 			.createQueryBuilder('label')
-			.where(andWhere([serviceCondition, userCondition]), { serviceId, ...userParams })
+			.where(andWhere([serviceCondition, categoryCondition, userCondition]), {
+				serviceIds,
+				categoryIds,
+				...userParams,
+			})
 			.leftJoin('label.service', 'servicelabel')
+			.leftJoin('label.category', 'categorylabel')
 			.getMany();
+	}
+
+	public async populateLabelForCategories<T extends LabelCategory>(entries: T[]): Promise<T[]> {
+		const ids = entries.map((s) => {
+			return s.id;
+		});
+		const listLabels = await this.find({ categoryIds: ids, skipAuthorisation: true });
+		const lookupTable = groupByKey(listLabels, (b) => b.categoryId);
+		entries.forEach((entry) => entry.labels = lookupTable.get(entry.id) || []);
+
+		return entries;
+	}
+
+	public async populateLabelForService<T extends Service>(entries: T[]): Promise<T[]> {
+		const ids = entries.map((s) => {
+			return s.id;
+		});
+		const listLabels = await this.find({ serviceIds: ids });
+		const lookupTable = groupByKey(listLabels, (b) => b.serviceId);
+		entries.forEach((entry) => entry.labels = lookupTable.get(entry.id) || []);
+
+		return entries;
 	}
 }
